Labfiles/04-semantic-kernel/python/requirements.txt
openai
python-dotenv
semantic_kernel